+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -137,7 +137,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-09-10T15:38:22.518Z")
+    updated: Optional[datetime] = parse_datetime("2026-09-11T00:37:36.779Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")